Updating how the integration setup script find the latest version of the packaged Web SDK module.
We were using npm view to get the latest version of this package, but
unfortunately that only checks released packages and not the local one.
Switching to a more generic method of finding the local tar file for the latest
version of the SDK, this should work fine on the CI build and locally when
building multiple versions.