Scrap version of the OAuth post message option.
We may want to do something like this to ease clients who are in hybrid
scenarios.

This way, their native app will get the redirect, and they will need to
find a way to post message that redirect to the window in the webview.

The sdk will listen for a particular message that has the redirect,
parse it, then send the expected post message to the widget so it can
respond.